Move domain participant entity into its own module
