+++ dds/src/dcps/dcps_domain_participant/builtin_publisher.rs
@@ -18,7 +18,7 @@ use crate::{
 };
 use alloc::string::String;
 
-use super::{
+use super::builtin_constants::{
     ENTITYID_BUILTIN_PUBLISHER, ENTITYID_SEDP_BUILTIN_PUBLICATIONS_ANNOUNCER,
     ENTITYID_SEDP_BUILTIN_SUBSCRIPTIONS_ANNOUNCER, ENTITYID_SEDP_BUILTIN_TOPICS_ANNOUNCER,
     ENTITYID_SPDP_BUILTIN_PARTICIPANT_WRITER, ENTITYID_TL_SVC_REPLY_WRITER,
