+++ dds/src/dcps/dcps_domain_participant/builtin_subscriber.rs
@@ -10,7 +10,7 @@ use crate::{
 use alloc::string::String;
 use core::ops::{Deref, DerefMut};
 
-use super::{
+use super::builtin_constants::{
     ENTITYID_BUILTIN_SUBSCRIBER, ENTITYID_SEDP_BUILTIN_PUBLICATIONS_DETECTOR,
     ENTITYID_SEDP_BUILTIN_SUBSCRIPTIONS_DETECTOR, ENTITYID_SEDP_BUILTIN_TOPICS_DETECTOR,
     ENTITYID_SPDP_BUILTIN_PARTICIPANT_READER, ENTITYID_TL_SVC_REPLY_READER,
