+++ dds/src/dcps/dcps_domain_participant/participant_entity.rs
@@ -1,0 +1,553 @@
+use super::{
+    builtin_constants::{
+        ENTITYID_SEDP_PUBLICATIONS_TOPIC, ENTITYID_SEDP_SUBSCRIPTIONS_TOPIC,
+        ENTITYID_SEDP_TOPICS_TOPIC, ENTITYID_SPDP_TOPIC, ENTITYID_TL_SVC_REQ_TOPIC,
+        ENTITYID_TL_SVC_RPL_TOPIC, TYPE_LOOKUP_REPLY_TOPIC_NAME, TYPE_LOOKUP_REQUEST_TOPIC_NAME,
+    },
+    builtin_publisher::BuiltinPublisher,
+    builtin_subscriber::BuiltinSubscriber,
+    topic_entity::{ContentFilteredTopicEntity, TopicEntity},
+    user_defined_publisher::PublisherEntity,
+    user_defined_subscriber::UserDefinedSubscriber,
+};
+use crate::{
+    builtin_topics::{
+        BuiltInTopicKey, DCPS_PARTICIPANT, DCPS_PUBLICATION, DCPS_SUBSCRIPTION, DCPS_TOPIC,
+        ParticipantBuiltinTopicData, PublicationBuiltinTopicData, SubscriptionBuiltinTopicData,
+        TopicBuiltinTopicData,
+    },
+    dcps::{
+        channels::{mpsc::MpscSender, oneshot::OneshotSender},
+        data_representation_builtin_endpoints::{
+            discovered_reader_data::DiscoveredReaderData,
+            discovered_writer_data::DiscoveredWriterData,
+            type_lookup::{TypeLookupReply, TypeLookupRequest},
+        },
+        listeners::domain_participant_listener::ListenerMail,
+        status_condition::DcpsStatusCondition,
+        status_mask::StatusMask,
+    },
+    dds_async::domain_participant_factory::DcpsSender,
+    infrastructure::{
+        domain::DomainId,
+        error::{DdsError, DdsResult},
+        instance::InstanceHandle,
+        qos::{DomainParticipantQos, PublisherQos, SubscriberQos, TopicQos},
+        time::{Duration, DurationKind, Time},
+    },
+    runtime::Timer,
+    transport::{
+        interface::RtpsTransportParticipant,
+        types::{ENTITYID_PARTICIPANT, Guid, GuidPrefix, Locator, USER_DEFINED_TOPIC},
+    },
+    xtypes::{
+        dynamic_type::DynamicType,
+        type_support::{Type, TypeSupport},
+    },
+};
+use alloc::{
+    boxed::Box,
+    collections::BTreeSet,
+    string::{String, ToString},
+    vec::Vec,
+};
+use core::{
+    future::{Future, poll_fn},
+    pin::{Pin, pin},
+    task::Poll,
+};
+
+pub struct DiscoveredParticipantInfo {
+    pub dds_participant_data: ParticipantBuiltinTopicData,
+    pub guid_prefix: GuidPrefix,
+    pub default_unicast_locator_list: Vec<Locator>,
+    pub default_multicast_locator_list: Vec<Locator>,
+    pub lease_duration: Duration,
+    pub last_communication_timestamp: Time,
+}
+
+pub fn poll_timeout<T>(
+    mut timer_handle: impl Timer,
+    duration: core::time::Duration,
+    mut future: Pin<Box<dyn Future<Output = T> + Send>>,
+) -> impl Future<Output = DdsResult<T>> {
+    poll_fn(move |cx| {
+        let timeout = timer_handle.delay(duration);
+        if let Poll::Ready(t) = pin!(&mut future).poll(cx) {
+            return Poll::Ready(Ok(t));
+        }
+        if pin!(timeout).poll(cx).is_ready() {
+            return Poll::Ready(Err(DdsError::Timeout));
+        }
+
+        Poll::Pending
+    })
+}
+
+#[derive(Debug, Clone, TypeSupport)]
+pub struct BuiltInKeyHolder {
+    #[dust_dds(key)]
+    pub key: BuiltInTopicKey,
+}
+
+pub struct DcpsDomainParticipant {
+    pub transport: RtpsTransportParticipant,
+
+    pub reader_counter: u16,
+    pub writer_counter: u16,
+    pub publisher_counter: u8,
+    pub subscriber_counter: u8,
+    pub domain_participant: DomainParticipantEntity,
+    pub dcps_sender: DcpsSender,
+}
+
+impl DcpsDomainParticipant {
+    #[allow(clippy::too_many_arguments)]
+    pub fn new(
+        domain_id: DomainId,
+        domain_tag: String,
+        guid_prefix: GuidPrefix,
+        domain_participant_qos: DomainParticipantQos,
+        listener_sender: Option<MpscSender<ListenerMail>>,
+        listener_mask: StatusMask,
+        transport: RtpsTransportParticipant,
+        dcps_sender: DcpsSender,
+    ) -> Self {
+        let guid = Guid::new(guid_prefix, ENTITYID_PARTICIPANT);
+
+        let participant_handle = InstanceHandle::new(guid.into());
+
+        const NUMBER_BUILTIN_ENTITIES: usize = 6;
+        let mut topic_list = Vec::with_capacity(NUMBER_BUILTIN_ENTITIES);
+
+        topic_list.push(TopicEntity::new(
+            TopicQos::default(),
+            "SpdpDiscoveredParticipantData".to_string(),
+            String::from(DCPS_PARTICIPANT),
+            InstanceHandle::new(Guid::new(guid_prefix, ENTITYID_SPDP_TOPIC).into()),
+            DcpsStatusCondition::default(),
+            None,
+            StatusMask::default(),
+            ParticipantBuiltinTopicData::TYPE,
+        ));
+
+        topic_list.push(TopicEntity::new(
+            TopicQos::default(),
+            "DiscoveredTopicData".to_string(),
+            String::from(DCPS_TOPIC),
+            InstanceHandle::new(Guid::new(guid_prefix, ENTITYID_SEDP_TOPICS_TOPIC).into()),
+            DcpsStatusCondition::default(),
+            None,
+            StatusMask::default(),
+            TopicBuiltinTopicData::TYPE,
+        ));
+
+        topic_list.push(TopicEntity::new(
+            TopicQos::default(),
+            "DiscoveredWriterData".to_string(),
+            String::from(DCPS_PUBLICATION),
+            InstanceHandle::new(Guid::new(guid_prefix, ENTITYID_SEDP_PUBLICATIONS_TOPIC).into()),
+            DcpsStatusCondition::default(),
+            None,
+            StatusMask::default(),
+            PublicationBuiltinTopicData::TYPE,
+        ));
+
+        topic_list.push(TopicEntity::new(
+            TopicQos::default(),
+            "DiscoveredReaderData".to_string(),
+            String::from(DCPS_SUBSCRIPTION),
+            InstanceHandle::new(Guid::new(guid_prefix, ENTITYID_SEDP_SUBSCRIPTIONS_TOPIC).into()),
+            DcpsStatusCondition::default(),
+            None,
+            StatusMask::default(),
+            SubscriptionBuiltinTopicData::TYPE,
+        ));
+
+        topic_list.push(TopicEntity::new(
+            TopicQos::default(),
+            "TypeLookup_Request".to_string(),
+            String::from(TYPE_LOOKUP_REQUEST_TOPIC_NAME),
+            InstanceHandle::new(Guid::new(guid_prefix, ENTITYID_TL_SVC_REQ_TOPIC).into()),
+            DcpsStatusCondition::default(),
+            None,
+            StatusMask::default(),
+            TypeLookupRequest::TYPE,
+        ));
+
+        topic_list.push(TopicEntity::new(
+            TopicQos::default(),
+            "TypeLookup_Reply".to_string(),
+            String::from(TYPE_LOOKUP_REPLY_TOPIC_NAME),
+            InstanceHandle::new(Guid::new(guid_prefix, ENTITYID_TL_SVC_RPL_TOPIC).into()),
+            DcpsStatusCondition::default(),
+            None,
+            StatusMask::default(),
+            TypeLookupReply::TYPE,
+        ));
+
+        let builtin_subscriber = BuiltinSubscriber::new(guid_prefix);
+        let builtin_publisher = BuiltinPublisher::new(guid_prefix, &transport);
+
+        let domain_participant = DomainParticipantEntity::new(
+            domain_id,
+            domain_participant_qos,
+            listener_sender,
+            listener_mask,
+            participant_handle,
+            builtin_publisher,
+            builtin_subscriber,
+            topic_list,
+            domain_tag,
+        );
+
+        Self {
+            transport,
+            reader_counter: 0,
+            writer_counter: 0,
+            publisher_counter: 0,
+            subscriber_counter: 0,
+            domain_participant,
+            dcps_sender,
+        }
+    }
+
+    pub fn domain_id(&self) -> DomainId {
+        self.domain_participant.domain_id
+    }
+
+    pub fn time_until_stale_participant(&self, now: Time) -> Option<Duration> {
+        self.domain_participant
+            .discovered_participant_list
+            .iter()
+            .map(|dp| dp.lease_duration - (now - dp.last_communication_timestamp))
+            .min()
+    }
+
+    pub fn time_until_missed_reader_deadline(&self, now: Time) -> Option<Duration> {
+        self.domain_participant
+            .user_defined_subscriber_list
+            .iter()
+            .flat_map(|subscriber| subscriber.data_reader_list.iter())
+            .filter_map(|data_reader| {
+                if let DurationKind::Finite(deadline) = data_reader.qos.deadline.period {
+                    data_reader
+                        .instance_ownership
+                        .iter()
+                        .map(|instance| deadline - (now - instance.last_received_time))
+                        .min()
+                } else {
+                    None
+                }
+            })
+            .min()
+    }
+
+    pub fn time_until_missed_writer_deadline(&self, now: Time) -> Option<Duration> {
+        self.domain_participant
+            .user_defined_publisher_list
+            .iter()
+            .flat_map(|publisher| publisher.data_writer_list.iter())
+            .filter_map(|data_writer| {
+                if let DurationKind::Finite(deadline) = data_writer.qos.deadline.period {
+                    data_writer
+                        .registered_instance_info
+                        .iter()
+                        .filter_map(|instance| instance.last_write_time)
+                        .map(|last_write_time| deadline - (now - last_write_time))
+                        .min()
+                } else {
+                    None
+                }
+            })
+            .min()
+    }
+
+    pub fn time_until_stale_writer_sample(&self, now: Time) -> Option<Duration> {
+        self.domain_participant
+            .user_defined_publisher_list
+            .iter()
+            .flat_map(|publisher| publisher.data_writer_list.iter())
+            .filter_map(|data_writer| {
+                if let DurationKind::Finite(lifespan) = data_writer.qos.lifespan.duration {
+                    data_writer
+                        .transport_writer
+                        .changes()
+                        .iter()
+                        .filter_map(|cc| cc.source_timestamp)
+                        .map(|source_timestamp| Time::from(source_timestamp) + lifespan - now)
+                        .min()
+                } else {
+                    None
+                }
+            })
+            .min()
+    }
+
+    pub fn get_instance_handle(&self) -> &InstanceHandle {
+        &self.domain_participant.instance_handle
+    }
+}
+
+pub const BUILT_IN_TOPIC_NAME_LIST: [&str; 6] = [
+    DCPS_PARTICIPANT,
+    DCPS_TOPIC,
+    DCPS_PUBLICATION,
+    DCPS_SUBSCRIPTION,
+    TYPE_LOOKUP_REQUEST_TOPIC_NAME,
+    TYPE_LOOKUP_REPLY_TOPIC_NAME,
+];
+
+pub struct FindTopicNotification {
+    pub topic_name: String,
+    pub deadline: Time,
+    pub type_support: DynamicType<'static>,
+    pub reply_sender: OneshotSender<DdsResult<(InstanceHandle, String)>>,
+}
+
+pub struct DomainParticipantEntity {
+    pub domain_id: DomainId,
+    pub domain_tag: String,
+    pub topic_counter: u16,
+    pub instance_handle: InstanceHandle,
+    pub qos: DomainParticipantQos,
+    pub builtin_subscriber: BuiltinSubscriber,
+    pub builtin_publisher: BuiltinPublisher,
+    pub user_defined_subscriber_list: Vec<UserDefinedSubscriber>,
+    pub default_subscriber_qos: SubscriberQos,
+    pub user_defined_publisher_list: Vec<PublisherEntity>,
+    pub default_publisher_qos: PublisherQos,
+    pub locally_created_topic_list: Vec<TopicEntity>,
+    pub content_filtered_topic_list: Vec<ContentFilteredTopicEntity>,
+    pub default_topic_qos: TopicQos,
+    pub discovered_participant_list: Vec<DiscoveredParticipantInfo>,
+    pub discovered_topic_list: Vec<TopicBuiltinTopicData>,
+    pub discovered_reader_list: Vec<DiscoveredReaderData>,
+    pub discovered_writer_list: Vec<DiscoveredWriterData>,
+    pub enabled: bool,
+    pub ignored_participants: BTreeSet<InstanceHandle>,
+    pub ignored_publications: BTreeSet<InstanceHandle>,
+    pub ignored_subscriptions: BTreeSet<InstanceHandle>,
+    pub _ignored_topic_list: BTreeSet<InstanceHandle>,
+    pub listener_sender: Option<MpscSender<ListenerMail>>,
+    pub listener_mask: StatusMask,
+    pub find_topic_sender_list: Vec<FindTopicNotification>,
+}
+
+impl DomainParticipantEntity {
+    #[allow(clippy::too_many_arguments)]
+    pub const fn new(
+        domain_id: DomainId,
+        domain_participant_qos: DomainParticipantQos,
+        listener_sender: Option<MpscSender<ListenerMail>>,
+        listener_mask: StatusMask,
+        instance_handle: InstanceHandle,
+        builtin_publisher: BuiltinPublisher,
+        builtin_subscriber: BuiltinSubscriber,
+        locally_created_topic_list: Vec<TopicEntity>,
+        domain_tag: String,
+    ) -> Self {
+        Self {
+            domain_id,
+            instance_handle,
+            topic_counter: 0,
+            qos: domain_participant_qos,
+            builtin_subscriber,
+            builtin_publisher,
+            user_defined_subscriber_list: Vec::new(),
+            default_subscriber_qos: SubscriberQos::const_default(),
+            user_defined_publisher_list: Vec::new(),
+            default_publisher_qos: PublisherQos::const_default(),
+            locally_created_topic_list,
+            content_filtered_topic_list: Vec::new(),
+            default_topic_qos: TopicQos::const_default(),
+            discovered_participant_list: Vec::new(),
+            discovered_topic_list: Vec::new(),
+            discovered_reader_list: Vec::new(),
+            discovered_writer_list: Vec::new(),
+            enabled: false,
+            ignored_participants: BTreeSet::new(),
+            ignored_publications: BTreeSet::new(),
+            ignored_subscriptions: BTreeSet::new(),
+            _ignored_topic_list: BTreeSet::new(),
+            listener_sender,
+            listener_mask,
+            domain_tag,
+            find_topic_sender_list: Vec::new(),
+        }
+    }
+
+    pub fn add_discovered_topic(&mut self, topic_builtin_topic_data: TopicBuiltinTopicData) {
+        match self
+            .discovered_topic_list
+            .iter_mut()
+            .find(|t| t.key() == topic_builtin_topic_data.key())
+        {
+            Some(x) => *x = topic_builtin_topic_data,
+            None => self.discovered_topic_list.push(topic_builtin_topic_data),
+        }
+    }
+
+    pub fn remove_discovered_writer(&mut self, discovered_writer_handle: &InstanceHandle) {
+        self.discovered_writer_list
+            .retain(|x| &x.dds_publication_data.key().value != discovered_writer_handle.as_ref());
+    }
+
+    pub fn get_discovered_topic_data(
+        &self,
+        topic_handle: &InstanceHandle,
+    ) -> Option<&TopicBuiltinTopicData> {
+        self.discovered_topic_list
+            .iter()
+            .find(|x| &x.key().value == topic_handle.as_ref())
+    }
+
+    pub fn find_topic(
+        &mut self,
+        topic_name: &str,
+        type_support: DynamicType<'static>,
+    ) -> Option<(InstanceHandle, String)> {
+        if let Some(topic) = self
+            .locally_created_topic_list
+            .iter()
+            .find(|x| x.topic_name == topic_name)
+        {
+            Some((topic.instance_handle, topic.type_name.clone()))
+        } else if let Some(discovered_topic_data) = self
+            .discovered_topic_list
+            .iter()
+            .find(|&discovered_topic_data| discovered_topic_data.name() == topic_name)
+        {
+            let qos = TopicQos {
+                topic_data: discovered_topic_data.topic_data().clone(),
+                durability: discovered_topic_data.durability().clone(),
+                deadline: discovered_topic_data.deadline().clone(),
+                latency_budget: discovered_topic_data.latency_budget().clone(),
+                liveliness: discovered_topic_data.liveliness().clone(),
+                reliability: discovered_topic_data.reliability().clone(),
+                destination_order: discovered_topic_data.destination_order().clone(),
+                history: discovered_topic_data.history().clone(),
+                resource_limits: discovered_topic_data.resource_limits().clone(),
+                transport_priority: discovered_topic_data.transport_priority().clone(),
+                lifespan: discovered_topic_data.lifespan().clone(),
+                ownership: discovered_topic_data.ownership().clone(),
+                representation: discovered_topic_data.representation().clone(),
+            };
+            let type_name = discovered_topic_data.type_name.clone();
+            let topic_handle = InstanceHandle::new([
+                self.instance_handle[0],
+                self.instance_handle[1],
+                self.instance_handle[2],
+                self.instance_handle[3],
+                self.instance_handle[4],
+                self.instance_handle[5],
+                self.instance_handle[6],
+                self.instance_handle[7],
+                self.instance_handle[8],
+                self.instance_handle[9],
+                self.instance_handle[10],
+                self.instance_handle[11],
+                0,
+                self.topic_counter.to_ne_bytes()[0],
+                self.topic_counter.to_ne_bytes()[1],
+                USER_DEFINED_TOPIC,
+            ]);
+            self.topic_counter += 1;
+            let status_condition = DcpsStatusCondition::default();
+            let mut topic = TopicEntity::new(
+                qos,
+                type_name.clone().into(),
+                String::from(topic_name),
+                topic_handle,
+                status_condition,
+                None,
+                StatusMask::default(),
+                type_support,
+            );
+            topic.enabled = true;
+
+            match self
+                .locally_created_topic_list
+                .iter_mut()
+                .find(|x| x.topic_name == topic.topic_name)
+            {
+                Some(x) => *x = topic,
+                None => self.locally_created_topic_list.push(topic),
+            }
+
+            Some((topic_handle, type_name.into()))
+        } else {
+            None
+        }
+    }
+
+    pub fn add_discovered_reader(&mut self, discovered_reader_data: DiscoveredReaderData) {
+        match self.discovered_reader_list.iter_mut().find(|x| {
+            x.dds_subscription_data.key() == discovered_reader_data.dds_subscription_data.key()
+        }) {
+            Some(x) => *x = discovered_reader_data,
+            None => self.discovered_reader_list.push(discovered_reader_data),
+        }
+    }
+
+    pub fn remove_discovered_reader(&mut self, discovered_reader_handle: &InstanceHandle) {
+        self.discovered_reader_list
+            .retain(|x| &x.dds_subscription_data.key().value != discovered_reader_handle.as_ref());
+    }
+
+    pub fn add_discovered_writer(&mut self, discovered_writer_data: DiscoveredWriterData) {
+        match self.discovered_writer_list.iter_mut().find(|x| {
+            x.dds_publication_data.key() == discovered_writer_data.dds_publication_data.key()
+        }) {
+            Some(x) => *x = discovered_writer_data,
+            None => self.discovered_writer_list.push(discovered_writer_data),
+        }
+    }
+
+    pub fn remove_subscriber(&mut self, handle: &InstanceHandle) -> Option<UserDefinedSubscriber> {
+        let i = self
+            .user_defined_subscriber_list
+            .iter()
+            .position(|x| &x.instance_handle == handle)?;
+
+        Some(self.user_defined_subscriber_list.remove(i))
+    }
+
+    pub fn remove_publisher(&mut self, handle: &InstanceHandle) -> Option<PublisherEntity> {
+        let i = self
+            .user_defined_publisher_list
+            .iter()
+            .position(|x| &x.instance_handle == handle)?;
+
+        Some(self.user_defined_publisher_list.remove(i))
+    }
+
+    pub fn is_empty(&self) -> bool {
+        let no_user_defined_topics = self
+            .locally_created_topic_list
+            .iter()
+            .filter(|t| !BUILT_IN_TOPIC_NAME_LIST.contains(&t.topic_name.as_str()))
+            .count()
+            == 0;
+
+        self.user_defined_publisher_list.is_empty()
+            && self.user_defined_subscriber_list.is_empty()
+            && self.content_filtered_topic_list.is_empty()
+            && no_user_defined_topics
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_time_until_stale_writer_sample_calculation() {
+        let source_timestamp = crate::transport::types::Time::new(100, 0);
+        let lifespan = Duration::new(10, 0);
+        let now = Time::new(105, 0);
+
+        let remaining = Time::from(source_timestamp) + lifespan - now;
+        assert_eq!(remaining, Duration::new(5, 0));
+    }
+}
