+++ dds/src/dcps/dcps_domain_participant/topic_entity.rs
@@ -1,0 +1,123 @@
+use super::builtin_constants::{TYPE_LOOKUP_REPLY_TOPIC_NAME, TYPE_LOOKUP_REQUEST_TOPIC_NAME};
+use crate::{
+    builtin_topics::{
+        DCPS_PARTICIPANT, DCPS_PUBLICATION, DCPS_SUBSCRIPTION, DCPS_TOPIC,
+        ParticipantBuiltinTopicData, PublicationBuiltinTopicData, SubscriptionBuiltinTopicData,
+        TopicBuiltinTopicData,
+    },
+    dcps::{
+        channels::mpsc::MpscSender,
+        data_representation_builtin_endpoints::type_lookup::{TypeLookupReply, TypeLookupRequest},
+        listeners::domain_participant_listener::ListenerMail,
+        status_condition::DcpsStatusCondition,
+        status_mask::StatusMask,
+    },
+    infrastructure::{instance::InstanceHandle, qos::TopicQos, status::InconsistentTopicStatus},
+    xtypes::{
+        dynamic_type::DynamicType,
+        type_object::{TypeInformation, TypeObject},
+        type_support::Type,
+    },
+};
+use alloc::{string::String, vec::Vec};
+
+pub struct ContentFilteredTopicEntity {
+    pub topic_name: String,
+    pub related_topic_name: String,
+    pub filter_expression: String,
+    pub expression_parameters: Vec<String>,
+}
+
+impl ContentFilteredTopicEntity {
+    pub fn new(
+        name: String,
+        related_topic_name: String,
+        filter_expression: String,
+        expression_parameters: Vec<String>,
+    ) -> Self {
+        Self {
+            topic_name: name,
+            related_topic_name,
+            filter_expression,
+            expression_parameters,
+        }
+    }
+}
+
+#[allow(clippy::large_enum_variant)]
+pub enum DiscoveredTypeRepresentationState {
+    Requested,
+    Discovered(TypeObject),
+}
+
+pub struct TopicEntity {
+    pub qos: TopicQos,
+    pub type_name: String,
+    pub topic_name: String,
+    pub instance_handle: InstanceHandle,
+    pub enabled: bool,
+    pub inconsistent_topic_status: InconsistentTopicStatus,
+    pub status_condition: DcpsStatusCondition,
+    pub listener_sender: Option<MpscSender<ListenerMail>>,
+    pub listener_mask: StatusMask,
+    pub type_support: DynamicType<'static>,
+    pub type_information: TypeInformation,
+    pub discovered_type_representation: Vec<(TypeInformation, DiscoveredTypeRepresentationState)>,
+}
+
+impl TopicEntity {
+    #[allow(clippy::too_many_arguments)]
+    pub fn new(
+        qos: TopicQos,
+        type_name: String,
+        topic_name: String,
+        instance_handle: InstanceHandle,
+        status_condition: DcpsStatusCondition,
+        listener_sender: Option<MpscSender<ListenerMail>>,
+        listener_mask: StatusMask,
+        type_support: DynamicType<'static>,
+    ) -> Self {
+        Self {
+            qos,
+            type_name,
+            topic_name,
+            instance_handle,
+            enabled: false,
+            inconsistent_topic_status: InconsistentTopicStatus::const_default(),
+            status_condition,
+            listener_sender,
+            listener_mask,
+            type_support,
+            type_information: TypeInformation::from(type_support),
+            discovered_type_representation: Vec::new(),
+        }
+    }
+}
+
+pub fn get_topic_type_support<'a>(
+    topic_name: &str,
+    content_filtered_topic_list: &[ContentFilteredTopicEntity],
+    locally_created_topic_list: &'a [TopicEntity],
+) -> Option<&'a DynamicType<'static>> {
+    let resolved_topic_name = if let Some(cf_topic) = content_filtered_topic_list
+        .iter()
+        .find(|t| t.topic_name == topic_name)
+    {
+        cf_topic.related_topic_name.as_str()
+    } else {
+        topic_name
+    };
+
+    match resolved_topic_name {
+        DCPS_PARTICIPANT => Some(&ParticipantBuiltinTopicData::TYPE),
+        DCPS_TOPIC => Some(&TopicBuiltinTopicData::TYPE),
+        DCPS_PUBLICATION => Some(&PublicationBuiltinTopicData::TYPE),
+        DCPS_SUBSCRIPTION => Some(&SubscriptionBuiltinTopicData::TYPE),
+        TYPE_LOOKUP_REQUEST_TOPIC_NAME => Some(&TypeLookupRequest::TYPE),
+        TYPE_LOOKUP_REPLY_TOPIC_NAME => Some(&TypeLookupReply::TYPE),
+        _ => locally_created_topic_list
+            .iter()
+            .find(|t| t.topic_name == resolved_topic_name)
+            .map(|t| &t.type_support),
+    }
+}
